requirements.txt
rembg
Pillow
tkinterdnd2
